mqtt/requirements.txt
psycopg2-binary
sqlalchemy==2.0.15
paho-mqtt
